Added mealController.js
mealController.js handles the logic for everything pertaining to meal plan data. As of now it successfully retrieves the data submitted from the createmealpage and appends the meal_id to the associated user_id in the userdb. Also added routes to mealController in app.js, and added necessary logic in userController to handle retrieving the session of the current logged user.